analysis-api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
PyniteFEA==2.0.2
numpy==1.26.4
